+++ resources/views/users/index.blade.php
@@ -64,7 +64,7 @@
                                 <td>{{@$user->roles[0]->name}}</td>
                                 <td style="width: 50px;">
                                     
-                                    {!! Form::open(['route'=>['users.destroy',$user], 'method'=>'DELETE', 'id'=>'form_eliminar'.$user->id]) !!}
+                                    {!! Form::open(['route'=>['users.destroy',$user], 'method'=>'DELETE', 'id'=>'form_eliminar'.$user->id, 'onSubmit'=>'return confirm("¿Esta seguro de eliminar el usuario?")']) !!}
 
                                         <a class="btn btn-info btn-circle btn-sm" href="{{route('users.edit', $user)}}" title="Editar">
                                             <i class="far fa-edit"></i>
